Rename supplier program functions
